Reworking multicard support
Instead of syncronously waiting for open and close to finish, this
version implements the card access with help of statemachines. Those
should finally help to get the code in a shape where we can better
cope with failing cards. Those can be recovered, hopefully transparently
for the running application. This version is not yet perfect and is also
not passing all of our tests. It is just a starting.